Run gofmt against CDI source files
This commit just uses the newly added `make format` option to format the
existing source files in the CDI project.  We may not want to merge this
until we're confident in our unit and functional tests.

Most of the changes at this point are limited to removing spaces for
tabs and making basic formatting changes so it's relatively safe.