Change restart policy on import pod
This is a simple change that modifies the importer pod restart policy
from Never to on Error.  In the case of the importer we don't have to
coordinate multiple pods, so it's pretty easy to just leverage the
existing workload constructs to deal with transient failures.

For now, a restart will result in a partial download being restarted,
going forward it would be good to add enhancements to track progress and
resume a faile download.