Add the ability to deploy CDI to external provider
This patch adds the ability to build and deploy CDI to an existing
Kubernetes cluster, along the same lines as the external provider
available in Kubevirt.

The goal is to be able to build and deploy from source all of the CDI
manifests and upload them a Kubernetes cluster.  This could be your own
on prem hardware, or GKE.

************* WIP ********************
This sets the foundation for cluster-up without breaking things, and is
part way there for cluster-sync